System F elaboration
I've implemented HM with System F elaboration.
The notion of a type schema goes away completely which is nice.
The implementation becomes a little more complicated in other way
because we're dealing with full System F types, but that's only natural

The implementation has a few dark corners: I'm guessing the terms that
I elaborate into are not completely well-formed. I'll have to add tests
to that eventually. Also, the generalization generates too many
quantified variables. It doesn't hurt but is not ideal.

I've also restructured a few things in the UnionFind implementation
as I used that as a template and I found a few things that could
be cleaned up.